✨ Add word and phrase translation tooltips with OpenAI support
