+++ package.json
@@ -17,6 +17,7 @@
     "lucide-react": "^0.523.0",
     "next": "15.3.4",
     "next-auth": "^4.24.11",
+    "openai": "^5.8.2",
     "prisma": "^6.10.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
